+++ src/dom/Attr.js
@@ -1,14 +1,21 @@
 import { Node } from './Node.js'
-import { html } from '../utils/namespaces.js'
-
 export class Attr extends Node {
   constructor (name, props, ns) {
     super(name, { nodeValue: '', ...props }, ns)
 
-    // Follow spec and lowercase nodeName for html
-    this.nodeName = ns === html ? name.toLowerCase() : name
+    // createAttribute() performs HTML lowercasing before construction;
+    // createAttributeNS() must preserve the supplied qualified name.
+    this.nodeName = name
     this.nodeType = Node.ATTRIBUTE_NODE
     this.ownerElement = null
+  }
+
+  get nodeValue () {
+    return this._nodeValue
+  }
+
+  set nodeValue (val) {
+    this._nodeValue = String(val)
   }
 
   get value () {